Experiment/requirements/requirements.txt
pandas==1.0.3
scikit_learn==0.21.2
selenium==3.141.0
beautifulsoup4==4.9.1
Flask==1.1.2
mecab-python3==0.996.5
nltk==3.5
numpy==1.19.0
pyenchant==3.1.1
requests==2.23.0
selenium==3.141.0
urllib3==1.25.8
webdriver-manager==3.2.2
